Rename: slice 파일 이름 수정 apiSlice의 경우 이름에 Api 추가 globalState의 경우 set 접두사 추가
