+++ src/features/artists/services/artistsTypeApiSlice.ts
