+++ src/features/artists/services/setArtistSlice.ts
