+++ src/features/artists/services/setArtistTypeSlice.ts
