+++ src/features/categories/modals/CategoryModal.tsx
@@ -10,12 +10,12 @@ import {
   toggleCategory,
   toggleEditCategory,
 } from '../../modal/manageModalSlice';
-import { selectEditCategorySlice } from '../services/editCategorySlice';
 import {
   useGetEventCategoryQuery,
   useAddEventCategoryMutation,
   useEditEventCategoryMutation,
-} from '../services/eventCategorySlice';
+} from '../services/categoryApiSlice';
+import { selectEditCategorySlice } from '../services/setCategorySlice';
 
 import {
   CategoryModalTitle,
