+++ src/features/categories/services/setCategorySlice.ts
