+++ src/pages/BookmarkShare/index.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'react-router-dom';
 
 import logo from '../../assets/icons/logo.svg';
 import staricon from '../../assets/icons/starIcon.svg';
-import { useGetShareBookmarkFolderQuery } from '../../features/bookmarks/services/bookmarkFolderSlice';
+import { useGetShareBookmarkFolderQuery } from '../../features/bookmarks/services/bookmarkFolderApiSlice';
 import { useRouter } from '../../hooks/useRouter';
 import { shareFolderEventsType } from '../../types/bookmarkFolderType';
