+++ src/pages/ManagePage/Category/CategoryListItem.tsx
@@ -2,8 +2,8 @@ import { useDispatch } from 'react-redux';
 
 import deleteIcon from '../../../assets/icons/delete.svg';
 import editIcon from '../../../assets/icons/edit.svg';
-import { editCategoryInfo } from '../../../features/categories/services/editCategorySlice';
-import { useDeleteEventCategoryMutation } from '../../../features/categories/services/eventCategorySlice';
+import { useDeleteEventCategoryMutation } from '../../../features/categories/services/categoryApiSlice';
+import { editCategoryInfo } from '../../../features/categories/services/setCategorySlice';
 import { toggleEditCategory } from '../../../features/modal/manageModalSlice';
 
 import * as S from './CategoryListItemStyle';
