+++ src/pages/MyPage/Bookmark/BookmarkFolders.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux';
 import plusicon from '../../../assets/icons/cross.svg';
 import editicon from '../../../assets/icons/editpencil.svg';
 import starticon from '../../../assets/icons/starIcon.svg';
-import { useGetBookmarkFoldersQuery } from '../../../features/bookmarks/services/bookmarkFolderSlice';
+import { useGetBookmarkFoldersQuery } from '../../../features/bookmarks/services/bookmarkFolderApiSlice';
 import { toggleAddBookmarkFolder } from '../../../features/modal/manageModalSlice';
 import { BookmarkFolderType } from '../../../types/bookmarkFolderType';
